+++ src/Dan.Plugin.Arbt/Models/BemanningUnit.cs
@@ -1,0 +1,9 @@
+using Newtonsoft.Json;
+public class BemanningUnit
+{
+    [JsonProperty("Organisasjonsnummer")]
+    public string Organisasjonsnummer { get; set; }
+
+    [JsonProperty("Godkjenningsstatus")]
+    public string Godkjenningsstatus { get; set; }
+}
